- cli - add cli-option `--mode-report` - fs - merge file browser.mjs into index.html - jslint - disable linting of embedded javascript in markdown-files
