+++ function.html
@@ -189,10 +189,6 @@ a:active {
 <p>JSLint is delivered as a set of files:</p>
 <center><table align="center">
 <tr><th>Filename</th><th>Content</th></tr>
-<tr>
-    <td><code>browser.mjs</code></td>
-<td>Browser based UI and <code>report</code> object, containing report
-    generator functions for HTML.</td></tr>
 <tr><td><code>function.html</code></td>
 <td>This page that you are looking at right now.</td></tr>
 <tr><td><code>help.html</code></td>
